admin and moderator
admin and moderator

+++ Applications/src/main/java/DbClasses/Admin.java
@@ -2,7 +2,7 @@ package DbClasses;
 
 public class Admin extends Moderator {
 
-	public Admin(String username, String password, int balance,int accessLevel) {
+	public Admin(String username, String password, double balance,int accessLevel) {
 		super(username, password,balance,accessLevel);
 		setAccessLevel(2);
 	}
