+++ Applications/src/main/java/DbClasses/Moderator.java
@@ -2,7 +2,7 @@ package DbClasses;
 
 public class Moderator extends User {
 
-	public Moderator(String username, String password, int balance,int accessLevel) {
+	public Moderator(String username, String password, double balance,int accessLevel) {
 		super(username, password,balance,accessLevel);
 		setAccessLevel(1);
 	}
